refactor: split the hooks into multiples files and put them in a hooks directory
