+++ src/components/AdvancedSearchForm/AdvancedSearchForm.jsx
@@ -1,11 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-// import { Tooltip } from 'flowbite-react';
+import { Tooltip } from 'flowbite-react';
 
 import FormField from './FormFields/FormField';
 import OperatorRequest from './OperatorRequest/OperatorRequest';
-import { useStateWithCallback } from '../../lib/hooks';
-import { Tooltip } from 'flowbite-react';
+import useStateWithCallback from '@/hooks/stateWithCallback';
 
 function AdvancedSearchForm ({ queryInputHandler }) {
   const initialFieldArray = [{ type: 'field' }];
