+++ src/components/DownloadButton/DownloadButton.jsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Tooltip } from 'flowbite-react';
 
-import { useResetForm } from '@/lib/hooks';
-import { buildFullApiUrl, sendDownloadApiRequest, sendSaveQIdApiRequest } from '../../lib/istexApi';
 import ModalDownloadRewiews from './ModalDownloadRewiews';
+import { buildFullApiUrl, sendDownloadApiRequest, sendSaveQIdApiRequest } from '../../lib/istexApi';
+import useResetForm from '@/hooks/resetForm';
 import { useEventEmitterContext } from '@/contexts/EventEmitterContext';
 import { useHistoryContext } from '@/contexts/HistoryContext';
